moved lander to separate objects directory
